# server/target/openapi.yaml
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "Control Finanzas API",
                 "description":  "API para cuentas y ahorro (iteraciÃ³n inicial)",
                 "version":  "0.1.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "paths":  {
...
                  "/api/v1/transactions":  {
                                               "get":  {
                                                           "tags":  [
                                                                        "transaction-controller"
                                                                    ],
                                                           "operationId":  "list",
                                                           "parameters":  [
                                                                              {
                                                                                  "name":  "arg0",
                                                                                  "in":  "query",
                                                                                  "required":  false,
                                                                                  "schema":  {
                                                                                                 "type":  "string"
                                                                                             }
                                                                              }
                                                                          ],
                                                           "responses":  {
                                                                             "200":  {
                                                                                         "description":  "OK",
                                                                                         "content":  {
                                                                                                         "*/*":  {
                                                                                                                     "schema":  {
                                                                                                                                    "type":  "array",
                                                                                                                                    "items":  {
                                                                                                                                                  "$ref":  "#/components/schemas/Transaction"
                                                                                                                                              }
                                                                                                                                }
                                                                                                                 }
                                                                                                     }
                                                                                     }
                                                                         }
                                                       },
                                               "post":  {
                                                            "tags":  [
                                                                         "transaction-controller"
                                                                     ],
                                                            "operationId":  "create",
                                                            "requestBody":  {
                                                                                "content":  {
                                                                                                "application/json":  {
                                                                                                                         "schema":  {
                                                                                                                                        "$ref":  "#/components/schemas/CreateBody"
                                                                                                                                    }
                                                                                                                     }
                                                                                            },
                                                                                "required":  true
                                                                            },
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "OK",
                                                                                          "content":  {
                                                                                                          "*/*":  {
                                                                                                                      "schema":  {
                                                                                                                                     "type":  "object"
                                                                                                                                 }
                                                                                                                  }
                                                                                                      }
                                                                                      }
                                                                          }
                                                        }
                                           },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "CreateBody":  {
                                                          "required":  [
                                                                           "accountId",
                                                                           "amount",
                                                                           "categoryId",
                                                                           "date",
                                                                           "type"
                                                                       ],
                                                          "type":  "object",
                                                          "properties":  {
                                                                             "type":  {
                                                                                          "type":  "string",
                                                                                          "enum":  [
                                                                                                       "INCOME",
                                                                                                       "EXPENSE"
                                                                                                   ]
                                                                                      },
                                                                             "accountId":  {
                                                                                               "type":  "integer",
                                                                                               "format":  "int64"
                                                                                           },
                                                                             "categoryId":  {
                                                                                                "type":  "integer",
                                                                                                "format":  "int64"
                                                                                            },
                                                                             "amount":  {
                                                                                            "type":  "number",
                                                                                            "format":  "double"
                                                                                        },
                                                                             "date":  {
                                                                                          "type":  "string"
                                                                                      },
                                                                             "description":  {
                                                                                                 "type":  "string"
                                                                                             }
                                                                         }
                                                      },
...
                                       "Account":  {
                                                       "type":  "object",
                                                       "properties":  {
                                                                          "id":  {
                                                                                     "type":  "integer",
                                                                                     "format":  "int64"
                                                                                 },
                                                                          "name":  {
                                                                                       "type":  "string"
                                                                                   },
                                                                          "initialBalance":  {
                                                                                                 "type":  "number"
                                                                                             },
                                                                          "createdAt":  {
                                                                                            "type":  "string",
                                                                                            "format":  "date-time"
                                                                                        },
                                                                          "type":  {
                                                                                       "type":  "string",
                                                                                       "enum":  [
                                                                                                    "OPERATIVA",
                                                                                                    "AHORRO"
                                                                                                ]
                                                                                   }
                                                                      }
                                                   },
                                       "Category":  {
                                                        "type":  "object",
                                                        "properties":  {
                                                                           "id":  {
                                                                                      "type":  "integer",
                                                                                      "format":  "int64"
                                                                                  },
                                                                           "name":  {
                                                                                        "type":  "string"
                                                                                    }
                                                                       }
                                                    },
                                       "Transaction":  {
                                                           "type":  "object",
                                                           "properties":  {
                                                                              "id":  {
                                                                                         "type":  "integer",
                                                                                         "format":  "int64"
                                                                                     },
                                                                              "type":  {
                                                                                           "type":  "string",
                                                                                           "enum":  [
                                                                                                        "INCOME",
                                                                                                        "EXPENSE"
                                                                                                    ]
                                                                                       },
                                                                              "account":  {
                                                                                              "$ref":  "#/components/schemas/Account"
                                                                                          },
                                                                              "category":  {
                                                                                               "$ref":  "#/components/schemas/Category"
                                                                                           },
                                                                              "amount":  {
                                                                                             "type":  "number"
                                                                                         },
                                                                              "date":  {
                                                                                           "type":  "string",
                                                                                           "format":  "date"
                                                                                       },
                                                                              "description":  {
                                                                                                  "type":  "string"
                                                                                              }
                                                                          }
                                                       },
...
                                   }
                   }
}